[REF] calendar: Move renderer in its own file
