+++ addons/calendar/views/calendar_templates.xml
@@ -4,6 +4,7 @@
         <xpath expr="." position="inside">
             <link rel="stylesheet" type="text/scss" href="/calendar/static/src/scss/calendar.scss"/>
             <script type="text/javascript" src="/calendar/static/src/js/base_calendar.js"></script>
+            <script type="text/javascript" src="/calendar/static/src/js/calendar_renderer.js"></script>
             <script type="text/javascript" src="/calendar/static/src/js/calendar_controller.js"></script>
             <script type="text/javascript" src="/calendar/static/src/js/calendar_model.js"></script>
             <script type="text/javascript" src="/calendar/static/src/js/calendar_view.js"></script>
